Making ProtoStreamMarshaller default marshaller and adding infinispan-remote-query-client as required dependency
