+++ extensions/infinispan-client/runtime/pom.xml
@@ -59,7 +59,6 @@
         <dependency>
             <groupId>org.infinispan</groupId>
             <artifactId>infinispan-remote-query-client</artifactId>
-            <optional>true</optional>
         </dependency>
 
         <dependency>
