+++ extensions/infinispan-client/runtime/src/main/java/org/jboss/shamrock/infinispan/client/runtime/InfinispanClientProducer.java
@@ -1,13 +1,18 @@
 package org.jboss.shamrock.infinispan.client.runtime;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.lang.annotation.Annotation;
 import java.util.Map;
 import java.util.Properties;
+import java.util.Scanner;
 import java.util.Set;
 
 import javax.annotation.PreDestroy;
 import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.spi.CreationalContext;
 import javax.enterprise.inject.Produces;
+import javax.enterprise.inject.spi.Bean;
 import javax.enterprise.inject.spi.BeanManager;
 import javax.enterprise.inject.spi.InjectionPoint;
 import javax.inject.Inject;
@@ -20,10 +25,15 @@ import org.infinispan.client.hotrod.configuration.ConfigurationBuilder;
 import org.infinispan.client.hotrod.impl.ConfigurationProperties;
 import org.infinispan.client.hotrod.logging.Log;
 import org.infinispan.client.hotrod.logging.LogFactory;
+import org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller;
 import org.infinispan.commons.marshall.Marshaller;
 import org.infinispan.commons.util.Util;
 import org.infinispan.counter.api.CounterManager;
+import org.infinispan.protostream.FileDescriptorSource;
+import org.infinispan.protostream.MessageMarshaller;
+import org.infinispan.protostream.SerializationContext;
 import org.infinispan.query.remote.client.ProtobufMetadataManagerConstants;
+import org.infinispan.query.remote.client.impl.MarshallerRegistration;
 
 /**
  * Produces a configured remote cache manager instance
@@ -32,7 +42,6 @@ import org.infinispan.query.remote.client.ProtobufMetadataManagerConstants;
 public class InfinispanClientProducer {
    private static final Log log = LogFactory.getLog(InfinispanClientProducer.class);
 
-   public static final String PROTOBUF_MARSHALLER_CLASS_NAME = "org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller";
    public static final String PROTOBUF_FILE_PREFIX = "infinispan.client.hotrod.protofile.";
 
     private Properties properties;
@@ -82,20 +91,35 @@ public class InfinispanClientProducer {
         if (marshallerClassName != null) {
             Class<?> marshallerClass = Class.forName(marshallerClassName);
             properties.put(ConfigurationProperties.MARSHALLER, Util.getInstance(marshallerClass));
+        } else {
+           // Default to proto stream marshaller if one is not provided
+           properties.put(ConfigurationProperties.MARSHALLER, new ProtoStreamMarshaller());
         }
     }
 
    /**
-    * Sets up additional properties for use when querying is available. Note this method should only be invoked if
-    * protobuf is available (ie. {@link #isProtoBufAvailable(Object)}).
+    * Sets up additional properties for use when proto stream marshaller is in use
     * @param properties the properties to be updated for querying
     */
-    public static void handleQueryRequirements(Properties properties) {
+    public static void handleProtoStreamRequirements(Properties properties) {
        // We only apply this if we are substrate in build time to apply to the properties
        // Note that the other half is done in QuerySubstitutions.SubstituteMarshallerRegistration class
        // Note that the registration of these files are done twice in normal VM mode
        // (once during init and once at runtime)
-       HandleProtostreamMarshaller.handleQueryRequirements(properties);
+       properties.put(InfinispanClientProducer.PROTOBUF_FILE_PREFIX + MarshallerRegistration.QUERY_PROTO_RES,
+             getContents(MarshallerRegistration.QUERY_PROTO_RES));
+       properties.put(InfinispanClientProducer.PROTOBUF_FILE_PREFIX + MarshallerRegistration.MESSAGE_PROTO_RES,
+             getContents(MarshallerRegistration.MESSAGE_PROTO_RES));
+    }
+
+   /**
+    * Reads all the contents of the file as a single string using default charset
+    * @param fileName file on class path to read contents of
+    * @return string containing the contents of the file
+    */
+    private static String getContents(String fileName) {
+        InputStream stream = InfinispanClientProducer.class.getResourceAsStream(fileName);
+        return new Scanner(stream).useDelimiter("\\A").next();
     }
 
    /**
@@ -109,26 +133,58 @@ public class InfinispanClientProducer {
         ConfigurationBuilder builder = new ConfigurationBuilder();
         Object marshallerInstance = properties.remove(ConfigurationProperties.MARSHALLER);
         if (marshallerInstance != null) {
-            injectProtoMarshallers(marshallerInstance, properties);
+           if (marshallerInstance instanceof ProtoStreamMarshaller) {
+              handleProtoStreamMarshaller(marshallerInstance, properties, beanManager);
+           }
             builder.marshaller((Marshaller) marshallerInstance);
         }
         builder.withProperties(properties);
         return builder;
     }
 
-    // NOTE: that this method is removed during substitution in graal if protostream is not in the class path
-    private void injectProtoMarshallers(Object marshallerInstance, Properties properties) {
-       if (isProtoBufAvailable(marshallerInstance)) {
-          HandleProtostreamMarshaller.handlePossibleMarshaller(marshallerInstance, properties, beanManager);
-       }
-    }
+   private static void handleProtoStreamMarshaller(Object marshallerInstance, Properties properties, BeanManager beanManager) {
+      ProtoStreamMarshaller marshaller = (ProtoStreamMarshaller) marshallerInstance;
+      SerializationContext serializationContext = marshaller.getSerializationContext();
 
-    public static boolean isProtoBufAvailable(Object marshallerInstance) {
-       // proto stream is optional and we can't do Class.forName at Runtime and we don't want to attempt to load
-       // it in this class so we check the name and if present invoke another class that handles it
-       return marshallerInstance != null &&
-             marshallerInstance.getClass().getName().equals(PROTOBUF_MARSHALLER_CLASS_NAME);
-    }
+      FileDescriptorSource fileDescriptorSource = null;
+      for (Map.Entry<Object, Object> property : properties.entrySet()) {
+         Object key = property.getKey();
+         if (key instanceof String) {
+            String keyString = (String) key;
+            if (keyString.startsWith(InfinispanClientProducer.PROTOBUF_FILE_PREFIX)) {
+               String fileName = keyString.substring(InfinispanClientProducer.PROTOBUF_FILE_PREFIX.length());
+               String fileContents = (String) property.getValue();
+               if (fileDescriptorSource == null) {
+                  fileDescriptorSource = new FileDescriptorSource();
+               }
+               fileDescriptorSource.addProtoFile(fileName, fileContents);
+            }
+         }
+      }
+
+      try {
+         if (fileDescriptorSource != null) {
+            serializationContext.registerProtoFiles(fileDescriptorSource);
+         }
+
+         Set<Bean<FileDescriptorSource>> protoFileBeans = (Set) beanManager.getBeans(FileDescriptorSource.class);
+         for (Bean<FileDescriptorSource> bean : protoFileBeans) {
+            CreationalContext<FileDescriptorSource> ctx = beanManager.createCreationalContext(bean);
+            FileDescriptorSource fds = (FileDescriptorSource) beanManager.getReference(bean, FileDescriptorSource.class, ctx);
+            serializationContext.registerProtoFiles(fds);
+         }
+      } catch (IOException e) {
+         // TODO: pick a better exception
+         throw new RuntimeException(e);
+      }
+
+      Set<Bean<MessageMarshaller>> beans = (Set) beanManager.getBeans(MessageMarshaller.class);
+      for (Bean<MessageMarshaller> bean : beans) {
+         CreationalContext<MessageMarshaller> ctx = beanManager.createCreationalContext(bean);
+         MessageMarshaller messageMarshaller = (MessageMarshaller) beanManager.getReference(bean, MessageMarshaller.class, ctx);
+         serializationContext.registerMarshaller(messageMarshaller);
+      }
+   }
 
     @PreDestroy
     public void destroy() {
