+++ integration-tests/infinispan-client/src/main/resources/META-INF/hotrod-client.properties
@@ -1,6 +1,6 @@
 # infinispan.client.hotrod.server_list=localhost:11232
 # infinispan.client.hotrod.marshaller=org.infinispan.commons.marshall.UTF8StringMarshaller
-infinispan.client.hotrod.marshaller=org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller
+# infinispan.client.hotrod.marshaller=org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller
 infinispan.client.hotrod.near_cache.mode=INVALIDATED
 infinispan.client.hotrod.near_cache.max_entries=2
